# docs/swagger.yaml
basePath: /api/v1
...
host: leti-backend.onrender.com
info:
  contact: {}
  description: Leti Server REST API
  title: Leti API
  version: "1.0"
paths:
...
  /artisan-reviews/{id}/replies:
    get:
      description: Returns all replies on a specific review, ordered oldest first.
      parameters:
      - description: Review UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  type: object
                type: array
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get replies for a review
      tags:
      - Reviews
  /artisan-reviews/{id}/reply:
    post:
      consumes:
      - application/json
      description: The artisan can reply once to any review on their profile. The
        client can reply once after the artisan has responded. One reply per role
        per review.
      parameters:
      - description: Review UUID
        in: path
        name: id
        required: true
        type: string
      - description: Reply text
        in: body
        name: body
        required: true
        schema:
          properties:
            body:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reply to a review
      tags:
      - Reviews
  /artisan-reviews/{id}/review:
    post:
      consumes:
      - application/json
      description: A client can leave a rating and comment for an artisan after a
        completed and paid booking. Only one review per client per artisan — subsequent
        calls update the existing review.
      parameters:
      - description: Artisan UUID
        in: path
        name: id
        required: true
        type: string
      - description: rating 1–5
        in: body
        name: body
        required: true
        schema:
          properties:
            comment:
              type: string
            rating:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  avg_rating:
                    type: number
                  rating:
                    type: integer
                  review_id:
                    type: string
                  total_reviews:
                    type: integer
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Leave a review for an artisan
      tags:
      - Reviews
  /artisan-reviews/{id}/reviews:
    get:
      description: Returns paginated reviews for a specific artisan, including a rating
        summary breakdown and any replies on each review.
      parameters:
      - description: Artisan UUID
        in: path
        name: id
        required: true
        type: string
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  type: object
                type: array
              pagination:
                type: object
              status:
                type: string
              summary:
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get reviews for an artisan
      tags:
      - Reviews
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"